nats-queues/requirements.txt
debugpy>=1.5.1
nats-py>=2.1.0
asyncio>=3.4.3
